:sparkles: [iOS] 添加Url正则判断，修复下载页数据传递错误

+++ ios/DwebBrowser/DwebBrowser/Bridge/DownloadAppManager.swift
@@ -22,7 +22,7 @@ public class DownloadAppManager: NSObject {
         self.downloadView = controller.view
     }
     
-    @objc public func onListenProgress(progress: Float) {
+    @objc public func onListenProgress(progress: Double) {
         DispatchQueue.main.async {
             progressPublisher.send(progress)
         }
